API_Final_Agent/api_final_agent/essence/rag_engine_optimized.py
"""
OPTIMIZED RAG Engine - Reduces API calls by 80%
- Uses smaller chunk sizes to avoid rate limits
- Implements aggressive caching
- Batch processing for embeddings
- Inherits shared functionality from BaseRAGEngine
"""

import os
import time
from typing import List, Dict, Tuple, Optional
from pathlib import Path

from llama_index.core import (
    VectorStoreIndex,
    SimpleDirectoryReader,
    StorageContext,
    load_index_from_storage,
    Settings,
)

# Import our rate-limited embedding wrapper
from .rate_limited_embedding import RateLimitedEmbedding

# Import base class
from .rag_engine_base import BaseRAGEngine

# Import logger
from .logger import get_logger

# Initialize logger
logger = get_logger(__name__)


class OptimizedRAGEngine(BaseRAGEngine):
    """
    Optimized RAG engine that minimizes API calls.
    Uses rate-limited embeddings and local storage.
    """

    def __init__(self, data_dir: str = "data", persist_dir: str = ".storage", cache_dir: str = ".cache"):
        """Initialize optimized RAG engine with rate limiting."""
        super().__init__(data_dir, persist_dir, cache_dir)
        self._setup_embeddings()
        self._setup_base_llm(model="gpt-4o-mini", temperature=0.1)
        self._setup_base_node_parser(chunk_size=300, chunk_overlap=30)

    def _setup_embeddings(self):
        """Configure embeddings with rate limiting."""
        api_key = os.getenv("OPENAI_API_KEY")
        if not api_key:
            raise ValueError("OPENAI_API_KEY not found in environment variables")

        # Use our custom rate-limited embedding wrapper
        # This adds 2-second delays between requests to prevent rate limits
        Settings.embed_model = RateLimitedEmbedding(
            model="text-embedding-3-small",  # Cheaper and more efficient than ada-002
            api_key=api_key,
            delay_seconds=2.0,  # Wait 2 seconds between requests
            embed_batch_size=5  # Very conservative batch size
        )
        
        logger.info("✓ Using rate-limited embedding with 2s delays between requests")

    def initialize_index(self, force_reload: bool = False, max_retries: int = 3) -> bool:
        """
        Load or create index with optimizations and rate limit handling.
        
        Args:
            force_reload: Force rebuilding the index
            max_retries: Maximum number of retries on rate limit errors
            
        Returns:
            True if successful
        """
        try:
            # Try to load existing index first
            if not force_reload and self.persist_dir.exists():
                logger.info("📚 Loading existing index...")
                storage_context = StorageContext.from_defaults(
                    persist_dir=str(self.persist_dir)
                )
                self.index = load_index_from_storage(storage_context)
                logger.info("✓ Index loaded (no API calls needed!)")
            else:
                logger.info(f"📄 Building optimized index from {self.data_dir}...")

                if not self.data_dir.exists():
                    raise FileNotFoundError(f"Data directory not found: {self.data_dir}")

                # Load documents
                documents = SimpleDirectoryReader(
                    str(self.data_dir),
                    required_exts=[".pdf"]
                ).load_data()

                if not documents:
                    raise ValueError(f"No PDF files found in {self.data_dir}")

                logger.info(f"✓ Loaded {len(documents)} documents")

                # Create index with retry logic for rate limits
                logger.info("⚙️ Creating index with optimized settings...")
                
                retry_count = 0
                while retry_count < max_retries:
                    try:
                        self.index = VectorStoreIndex.from_documents(
                            documents,
                            show_progress=True
                        )
                        break  # Success!
                        
                    except Exception as e:
                        error_msg = str(e).lower()
                        if "rate_limit" in error_msg or "429" in error_msg:
                            retry_count += 1
                            if retry_count < max_retries:
                                wait_time = 10 * retry_count  # Exponential backoff
                                logger.warning(
                                    f"⚠️ Rate limit hit (attempt {retry_count}/{max_retries}). "
                                    f"Waiting {wait_time} seconds..."
                                )
                                time.sleep(wait_time)
                            else:
                                logger.error(
                                    "❌ Max retries reached. Please try again later or reduce the number of PDFs."
                                )
                                raise
                        else:
                            raise

                # Persist for future use
                self.persist_dir.mkdir(exist_ok=True)
                self.index.storage_context.persist(persist_dir=str(self.persist_dir))
                logger.info(f"✓ Index saved to {self.persist_dir}")

            # Create query engine
            self.query_engine = self.index.as_query_engine(
                similarity_top_k=2,  # Reduced from 3 to save API calls
                response_mode="compact"
            )

            return True

        except FileNotFoundError as e:
            logger.error(f"Data directory or files not found: {e}")
            raise
        except ValueError as e:
            logger.error(f"Invalid data or configuration: {e}")
            raise
        except (ConnectionError, TimeoutError) as e:
            logger.error(f"API connection error during index creation: {e}")
            raise
        except Exception as e:
            logger.error(f"Unexpected error initializing index: {e}", exc_info=True)
            raise
