API_Final_Agent/requirements.txt
# API_Final_Agent - Unified Service Dependencies
# Combines ACE_Framework and EssenceAI requirements

# Core LLM Providers (from ACE)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# EssenceAI Dependencies (optional - service works with mock if not installed)
# Uncomment to enable full EssenceAI functionality:
# llama-index>=0.10.0
# llama-index-llms-openai>=0.1.0
# llama-index-llms-anthropic>=0.1.0
# llama-index-embeddings-openai>=0.1.0
# weaviate-client>=3.25.0

# EssenceAI Data Processing
pandas>=2.0.0
plotly>=5.18.0
tavily-python>=0.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Testing (optional)
pytest>=8.0.0
pytest-asyncio>=0.21.0
